fix list item styles/assert correct id when deleting

+++ Employee.Web/src/app/employees/employees.service.ts
@@ -30,6 +30,10 @@ export class EmployeeService {
 	}
 
 	public deleteEmployee(employeeId: number) {
+		if (employeeId === 0) {
+			return;
+		}
+
 		return this.http.delete<EmployeeDetails>(EmployeeService.employeeUrl, { params: { id: employeeId } });
 	}
 }
